PC port foundation: CMake build, PLATFORM_PC conditionals, SDK stubs, runtime fixes, game data loading, bgfx rendering, GX state machine, SDL3 windowing, TEV shader pipeline, display list replay — full link + honest render milestones

+++ include/JSystem/JHostIO/JORMContext.h
@@ -191,7 +191,7 @@ public:
 
     void genNodeSub(const char* label, JORReflexible* i_node, u32, u32);
     void putNode(JORReflexible* obj) {
-        mOutputStream << (uintptr_t)obj;
+        mOutputStream << (u32)(uintptr_t)obj;
     }
     void invalidNode(JORReflexible* i_node, u32);
